fix(rules): add stricter reqs for prototype detection
Fixes #13

Before, if an Identifier was found it would go through the normal
flow and be analyzed. For example, `Object.keys` would go through.
The prototype check was done later and then the method name was
changed.

Now, the prototype detection looks for the Identifier and prototype
method name earlier to prevent leaks.