fix(rules): chaining expressions after get/set linted correctly
Fixes #26

Before, rule assumed form of `Error.stack` where `.stack` is a
member expression and only checked the get/set properties. In the
case of `Error.stack.slice()`, `.stack` becomes a member expression
and the rules only checks the prototype instead of the get/set
properties.

Now, Member expressions are checked for existence on prototype or
in get/set properties.